[BE] fix: 오늘의 복습문제 갯수 조회 SQL수정

+++ back/src/main/java/com/clover/recode/domain/statistics/service/StatisticsService.java
@@ -10,7 +10,7 @@ public interface StatisticsService {
 
     StatisticsListRes getStatisticsList(Authentication authentication);
 
-    Integer getReviewCnt(Authentication authentication);
+    Long getReviewCnt(Authentication authentication);
 
      List<TodayProblemDto> getReviews(Authentication authentication);
